backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
python-jose[cryptography]==3.3.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
itsdangerous==2.2.0
pydantic-settings==2.7.1
aiofiles==24.1.0
python-multipart==0.0.20
Pillow==11.1.0
slowapi==0.1.9
passlib[bcrypt]==1.7.4
bcrypt<4.1
